# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Passport API",
    "version": "0.1.0",
    "description": "Tamper-evident behavioral receipts for AI agents. Hash-only storage, ed25519 signatures, public verification."
  },
  "servers": [
    { "url": "https://passport.metis.gold", "description": "Production" }
  ],
  "paths": {
...
    "/api/v1/receipts": {
      "get": {
        "summary": "Search receipts (operator auth)",
        "parameters": [
          { "name": "domain", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "maximum": 100 } }
        ],
        "responses": { "200": { "description": "Receipt list" } }
      },
      "post": {
        "summary": "Issue a pending signed receipt",
        "responses": { "201": { "description": "Receipt issued" }, "401": { "description": "Unauthorized" } }
      }
    },
...
    "/api/v1/receipts/{id}/revoke": {
      "post": {
        "summary": "Revoke a receipt",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Receipt revoked" }, "409": { "description": "Already revoked" } }
      }
    },
...
    "/api/v1/passport/agents/{id}/evidence": {
      "get": {
        "summary": "Query evidence entries",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "source_type", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Evidence list" } }
      },
      "post": {
        "summary": "Ingest enrolled evidence",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Evidence ingested" } }
      }
    },
...
    "/api/v1/operator/api-keys/{keyHash}": {
      "delete": { "summary": "Delete API key", "parameters": [{ "name": "keyHash", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "Deleted" } } }
    },
    "/api/v1/operator/slashing-ledger": {
      "get": { "summary": "Slashing history", "parameters": [{ "name": "tranche", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Slashing entries" } } }
    },
...
    "/api/v1/webhooks/{id}": {
      "delete": { "summary": "Delete webhook subscription", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "Deleted" } } }
    },
...
  }
}